fix bag gtag
