+++ src/app/layout.jsx
@@ -7,25 +7,18 @@ import Footer from "@/libs/components/footer/Footer";
 
 import styles from "./page.module.scss";
 import Head from "next/head";
-import Script from "next/script";
 
 const montserrat = Montserrat({ subsets: ["cyrillic"] });
+
+import { Scripts } from "@/libs/components/scripts/Scripts";
 
 export default async function RootLayout({ children }) {
   return (
     <html lang="de-DE">
       <Head>
-        <Script>{`${(function (w, d, s, l, i) {
-          w[l] = w[l] || [];
-          w[l].push({ "gtm.start": new Date().getTime(), event: "gtm.js" });
-          var f = d.getElementsByTagName(s)[0],
-            j = d.createElement(s),
-            dl = l != "dataLayer" ? "&l=" + l : "";
-          j.async = true;
-          j.src = "https://www.googletagmanager.com/gtm.js?id=" + i + dl;
-          f.parentNode.insertBefore(j, f);
-        })(window, document, "script", "dataLayer", "GTM-TN2QSRR7")}`}</Script>
+        <Scripts />
       </Head>
+
       <body className={montserrat.className} suppressHydrationWarning={true}>
         <noscript>
           <iframe
