+++ src/libs/components/scripts/Scripts.jsx
@@ -1,0 +1,27 @@
+"use client";
+
+import Script from "next/script";
+import { useEffect, useState } from "react";
+
+const Scripts = () => {
+  const [isClient, setIsClient] = useState(false);
+  useEffect(() => setIsClient(true), []);
+  return (
+    isClient && (
+      <Script>
+        {`(function (w, d, s, l, i) {
+          w[l] = w[l] || [];
+          w[l].push({ "gtm.start": new Date().getTime(), event: "gtm.js" });
+          var f = d.getElementsByTagName(s)[0],
+            j = d.createElement(s),
+            dl = l != "dataLayer" ? "&l=" + l : "";
+          j.async = true;
+          j.src = "https://www.googletagmanager.com/gtm.js?id=" + i + dl;
+          f.parentNode.insertBefore(j, f);
+        })(window, document, "script", "dataLayer", "GTM-TN2QSRR7")`}
+      </Script>
+    )
+  );
+};
+
+export { Scripts };
